update GPL 384.10007: AiMesh: an innovative new router feature that connects multiple ASUS routers to create a whole-home WiFi network.

+++ release/src/router/rc/rc_ipsec.c
@@ -385,25 +385,70 @@ void ipsec_prof_fill(int prof_idx, char *p_data, ipsec_prof_type_t prof_type)
     return;
 }
 
+void ipsec_prof_fill_ext(int prof_idx, char *p_data, ipsec_prof_type_t prof_type)
+{
+	int i = 1;
+    char *p_end = NULL, *p_tmp = NULL, *ptr=NULL;;
+    p_end = p_data;
+
+    /*encryption_p1_ext*/
+    prof[prof_type][prof_idx].encryption_p1_ext = (uint8_t)ipsec_profile_int_parse(FLAG_NONE,
+                                                               p_end, &i);
+    p_end += i; /*to shifft next '>'*/
+
+	/*hash_p1_ext*/
+    prof[prof_type][prof_idx].hash_p1_ext = (uint8_t)ipsec_profile_int_parse(FLAG_NONE,
+                                                               p_end, &i);
+	p_end += i; /*to shifft next '>'*/
+
+	/*dh_group*/
+    prof[prof_type][prof_idx].dh_group = (uint8_t)ipsec_profile_int_parse(FLAG_NONE,
+                                                               p_end, &i);
+	p_end += i; /*to shifft next '>'*/
+
+	 /*encryption_p2_ext*/
+    prof[prof_type][prof_idx].encryption_p2_ext = (uint8_t)ipsec_profile_int_parse(FLAG_NONE,
+                                                               p_end, &i);
+    p_end += i; /*to shifft next '>'*/
+
+	/*hash_p2_ext*/
+	/*the last one doesn't need to parse ">".*/
+    prof[prof_type][prof_idx].hash_p2_ext = atoi(p_end);
+
+	/*the end of profile*/
+	return;
+}
+
 int pre_ipsec_prof_set()
 {
-    char buf[SZ_MIN];
+    char buf[SZ_MIN], buf_ext[SZ_MIN];
     char *p_tmp = NULL, buf1[SZ_BUF];
+	char *p_tmp_ext = NULL, buf1_ext[SZ_BUF];
     int i, rc = 0, prof_count = 0;
 
     p_tmp = &buf1[0];
+	p_tmp_ext = &buf1_ext[0];
     memset(p_tmp, 0, sizeof(char) * SZ_MIN);    
+	memset(p_tmp_ext, 0, sizeof(char) * SZ_MIN); 
 	for(prof_count = PROF_CLI; prof_count < PROF_ALL; prof_count++){
 	    for(i = 1; i <= MAX_PROF_NUM; i++){
-			if(PROF_SVR == prof_count)
+			if(PROF_SVR == prof_count){
 	        	sprintf(&buf[0], "ipsec_profile_%d", i);
-			else if(PROF_CLI == prof_count)
+				sprintf(&buf_ext[0], "ipsec_profile_%d_ext", i);
+			}
+			else if(PROF_CLI == prof_count){
 				sprintf(&buf[0], "ipsec_profile_client_%d", i);
-	        if(NULL != nvram_safe_get(&buf[0])){
+				sprintf(&buf_ext[0], "ipsec_profile_client_%d_ext", i);
+			}
+
+	        if(NULL != nvram_safe_get(&buf[0]) && NULL != nvram_safe_get(&buf_ext[0])){
 	            strcpy(p_tmp, nvram_safe_get(&buf[0]));
+				strcpy(p_tmp_ext, nvram_safe_get(&buf_ext[0]));
 	            /*to avoid nvram that it has not been inited ready*/
 	            if(0 != *p_tmp){
-		                ipsec_prof_fill(i-1, p_tmp,prof_count);
+		            ipsec_prof_fill(i-1, p_tmp,prof_count);
+					if(0 != *p_tmp_ext)
+						ipsec_prof_fill_ext(i-1, p_tmp_ext,prof_count);
 	                rc = 1;
 	            }
 	        }
@@ -489,9 +534,9 @@ void rc_strongswan_conf_set()
                 "  i_dont_care_about_security_and_use_aggressive_mode_psk = yes\n\n"
                 "  plugins {\n    include strongswan.d/charon/*.conf\n  }\n"
                 "  filelog {\n      /var/log/strongswan.charon.log {\n"
-                "        time_format = %%b %%e %%T\n        default = 1\n"
+                "        time_format = %%b %%e %%T\n        default = %d\n"
                 "        append = no\n        flush_line = yes\n"
-                "     }\n  }\n", nvram_get_int("ipsec_threads_num"));
+                "     }\n  }\n", nvram_get_int("ipsec_threads_num"), nvram_get_int("ipsec_log_level"));
     if(0 != rc){
         if(('n' != samba_prof.dns1[0]) && ('\0' != samba_prof.dns1[0])){
             fprintf(fp,"\n  dns1=%s\n", samba_prof.dns1);
@@ -1179,10 +1224,34 @@ void ipsec_conf_remote_set(FILE *fp, int prof_idx, ipsec_prof_type_t prof_type)
     return;
 }
 
+char* get_ike_esp_bit_convert( char *str, int maxNum, int n, int type)
+{
+	int i;
+	char tmpStr[12];
+	memset(str, 0, sizeof(str));
+	for(i = 0; i < maxNum; i++){
+		if((n >> i) & 0x1 ){
+			if(type == FLAG_IKE_ENCRYPT)
+				sprintf(tmpStr, "-%s", encryp[i]);
+			else if(type == FLAG_ESP_HASH)
+				sprintf(tmpStr, "-%s", hash[i]);
+			else if(type == FLAG_DH_GROUP)
+				sprintf(tmpStr, "-%s", dh_group[i]);
+			strcat(str, tmpStr);
+		}
+	}
+	return str;
+}
+
 void ipsec_conf_phase1_set(FILE *fp, int prof_idx, ipsec_prof_type_t prof_type)
 {
+	char str[128];
     fprintf(fp, "  ikelifetime=%d\n", prof[prof_type][prof_idx].keylife_p1);
-    if((ENCRYPTION_TYPE_MAX_NUM != prof[prof_type][prof_idx].encryption_p1) &&
+	fprintf(fp, "  ike=%s", get_ike_esp_bit_convert(str, ENCRYPTION_TYPE_MAX_NUM, prof[prof_type][prof_idx].encryption_p1_ext, FLAG_IKE_ENCRYPT) + 1);
+	fprintf(fp, "%s", get_ike_esp_bit_convert(str, HASH_TYPE_MAX_NUM, prof[prof_type][prof_idx].hash_p1_ext, FLAG_ESP_HASH));
+	fprintf(fp, "%s\n", get_ike_esp_bit_convert(str, DH_GROUP_MAX_NUM, prof[prof_type][prof_idx].dh_group, FLAG_DH_GROUP));
+
+    /*if((ENCRYPTION_TYPE_MAX_NUM != prof[prof_type][prof_idx].encryption_p1) &&
        (HASH_TYPE_MAX_NUM != prof[prof_type][prof_idx].hash_p1)){
         fprintf(fp,"  ike=%s-%s-%s\n", encryp[prof[prof_type][prof_idx].encryption_p1]
                   , hash[prof[prof_type][prof_idx].hash_p1], dh_group[DH_GROUP_14]);
@@ -1198,14 +1267,18 @@ void ipsec_conf_phase1_set(FILE *fp, int prof_idx, ipsec_prof_type_t prof_type)
                   , hash[HASH_TYPE_SHA1], dh_group[DH_GROUP_14]
                   , encryp[prof[prof_type][prof_idx].encryption_p1]
                   , hash[HASH_TYPE_SHA256], dh_group[DH_GROUP_14]);
-    }
+    }*/
     return;
 }
 
 void ipsec_conf_phase2_set(FILE *fp, int prof_idx, ipsec_prof_type_t prof_type)
 {
+	char str[128];
     fprintf(fp, "  keylife=%d\n", prof[prof_type][prof_idx].keylife_p2);
-    if((ENCRYPTION_TYPE_MAX_NUM != prof[prof_type][prof_idx].encryption_p2) && 
+	fprintf(fp, "  esp=%s", get_ike_esp_bit_convert(str, ENCRYPTION_TYPE_MAX_NUM, prof[prof_type][prof_idx].encryption_p2_ext, FLAG_IKE_ENCRYPT) + 1);
+	fprintf(fp, "%s\n", get_ike_esp_bit_convert(str, HASH_TYPE_MAX_NUM, prof[prof_type][prof_idx].hash_p2_ext, FLAG_ESP_HASH));
+
+    /*if((ENCRYPTION_TYPE_MAX_NUM != prof[prof_type][prof_idx].encryption_p2) && 
        (HASH_TYPE_MAX_NUM != prof[prof_type][prof_idx].hash_p2)){
         fprintf(fp,"  esp=%s-%s\n", encryp[prof[prof_type][prof_idx].encryption_p2]
                   , hash[prof[prof_type][prof_idx].hash_p2]);
@@ -1221,7 +1294,7 @@ void ipsec_conf_phase2_set(FILE *fp, int prof_idx, ipsec_prof_type_t prof_type)
                   , hash[HASH_TYPE_SHA1]
                   , encryp[prof[prof_type][prof_idx].encryption_p2]
                   , hash[HASH_TYPE_SHA256]);
-    }
+    }*/
     return;
 }
 void rc_ipsec_topology_set()
@@ -1313,26 +1386,54 @@ void rc_ipsec_topology_set()
     }
     return;
 }
+void rc_ipsec_nvram_convert_check()
+{
+	int i, rc = 0, prof_count = 0;
+	char buf[SZ_MIN], buf_ext[SZ_MIN];
+	char *nv=NULL, *nvp=NULL, *b=NULL;
+	for(prof_count = PROF_CLI; prof_count < PROF_ALL; prof_count++){
+	    for(i = 1; i <= MAX_PROF_NUM; i++){
+			if(PROF_SVR == prof_count){
+				sprintf(&buf[0], "ipsec_profile_%d", i);
+				sprintf(&buf_ext[0], "ipsec_profile_%d_ext", i);
+			}
+			else if(PROF_CLI == prof_count){
+				sprintf(&buf[0], "ipsec_profile_client_%d", i);
+				sprintf(&buf_ext[0], "ipsec_profile_client_%d_ext", i);
+			}
+			if(NULL != strcmp(nvram_safe_get(&buf[0]), "") && NULL == strcmp(nvram_safe_get(&buf_ext[0]),"")){
+				nv = nvp = strdup(nvram_safe_get(&buf[0]));
+				b = strsep(&nvp, ">");
+				if(NULL == strcmp(b, "4"))
+					nvram_set(&buf_ext[0], "0>0>0>0>0");	/* none */
+				else
+					nvram_set(&buf_ext[0], "6>6>255>6>6");	/* 3des-aes128-sha1-sha256-modp768-modp1024-modp1536-modp2048-modp3072-modp4096-modp6144-modp8192 */
+			}
 
+		}
+	}
+}
 void rc_ipsec_config_init()
 {
     memset((ipsec_samba_t *)&samba_prof, 0, sizeof(ipsec_samba_t));
     memset((ipsec_prof_t *)&prof[0][0], 0, sizeof(ipsec_prof_t) * MAX_PROF_NUM);
-    memset((pki_ca_t *)&ca_tab[0], 0, sizeof(pki_ca_t) * CA_FILES_MAX_NUM);
+    //memset((pki_ca_t *)&ca_tab[0], 0, sizeof(pki_ca_t) * CA_FILES_MAX_NUM);
 	memset((ipsec_samba_t *)&pre_samba_prof, 0, sizeof(ipsec_samba_t));
-    system("cp -rf /usr/etc/* /tmp/etc/");
-    mkdir("/jffs/ca_files", 0777);
+	if(!d_exists("/etc/ipsec.d") || !d_exists("/etc/strongswan.d"))
+		system("cp -rf /usr/etc/* /tmp/etc/");
+    //mkdir("/jffs/ca_files", 0777);
     /*ipsec.conf init*/    
     rc_ipsec_conf_default_init();
     rc_ipsec_psk_xauth_rw_init();
     /*ipsec.secrets init*/
-    rc_ipsec_set(IPSEC_INIT,PROF_ALL);
+    if(nvram_get_int("ipsec_server_enable") || nvram_get_int("ipsec_client_enable"))
+		rc_ipsec_set(IPSEC_INIT,PROF_ALL);
     //rc_ipsec_secrets_set();
     //rc_ipsec_conf_set();
     /*ipsec pki shell script default generate*/
     //rc_ipsec_ca_default_gen();
     //rc_ipsec_pki_gen_exec();
-    rc_ipsec_ca_init();
+    //rc_ipsec_ca_init();
     /*ca import*/
     //rc_ipsec_ca_import();
     return;
